requirements-prod.txt
# Production-only dependencies (create separate file)
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
langchain==0.3.0
langgraph==0.2.16
groq==0.11.0
openai==1.51.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.10
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
structlog==24.4.0
prometheus-client==0.21.0
orjson==3.10.7
asyncpg==0.29.0
click==8.1.7
python-slugify==8.0.4
email-validator==2.2.0
